Add incomplete REST-API backend
Moved webapi code to separate class, added implementation for test
datasource. Graphs can not be fetched with the REST-API backend.
